New features

+++ src/main/java/dev/luanc/library/repository/InfractionRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InfractionRepository extends JpaRepository<Infraction, Long> {
+public interface InfractionRepository extends JpaRepository<Infraction, Integer> {
 
     public Integer countInfractionByUser(User user);
 }
